Lazy-load mermaid
