+++ grafast/ruru-components/src/components/Mermaid.tsx
@@ -1,38 +1,18 @@
-import type { GrafastPlanJSON } from "grafast/mermaid";
-import { planToMermaid } from "grafast/mermaid";
+import type { GrafastPlanJSON } from "grafast";
 import type { FC } from "react";
-import { useEffect, useMemo } from "react";
-
-let initialized = false;
+import { lazy, Suspense } from "react";
+const MermaidReal = lazy(() => import("./MermaidReal.js"));
 
 export const Mermaid: FC<{ plan: GrafastPlanJSON }> = ({ plan }) => {
-  const diagram = useMemo(() => planToMermaid(plan), [plan]);
-  useEffect(() => {
-    if (window.mermaid) {
-      if (!initialized) {
-        initialized = true;
-        window.mermaid.initialize({
-          startOnLoad: true,
-          maxTextSize: 1000000,
-          maxEdges: 1000,
-        });
+  return (
+    <Suspense
+      fallback={
+        <div>
+          Mermaid hasn&apos;t (yet) loaded, so we cannot render plan diagrams
+        </div>
       }
-      if (diagram) {
-        window.mermaid.contentLoaded();
-      }
-    }
-  }, [diagram]);
-  if (window.mermaid) {
-    return (
-      <div className="mermaid" key={diagram}>
-        {diagram}
-      </div>
-    );
-  } else {
-    return (
-      <div>
-        Mermaid hasn&apos;t (yet) loaded, so we cannot render plan diagrams
-      </div>
-    );
-  }
+    >
+      <MermaidReal plan={plan} />
+    </Suspense>
+  );
 };
